Refactored VM and VMState
1. Added VMState forks
2. Moved VM.get_computation() to VMState.get_computation()
3. Moved `execute_transaction`, `make_receipt`, `validate_block` and `validate_uncle` from `VM` to `VMState`
4. Configured `opcodes` and `precompiles` in `VMState.__init__(...)`
5. Make sure that VMState only do read-only operations on the chaindb
