+++ evm/vm/__init__.py
@@ -4,6 +4,3 @@ from .base import (  # noqa: F401
 from .message import (  # noqa: F401
     Message,
 )
-from .vm_state import (  # noqa: F401
-    VMState,
-)
