+++ evm/vm/forks/byzantium/__init__.py
@@ -1,44 +1,17 @@
-from cytoolz import (
-    merge,
-)
-
-from evm import precompiles
 from evm.constants import MAX_UNCLE_DEPTH
-from evm.rlp.receipts import (
-    Receipt,
-)
-from evm.utils.address import (
-    force_bytes_to_address,
-)
 from evm.validation import (
     validate_lte,
 )
-
-from ..frontier import (
-    FRONTIER_PRECOMPILES,
-    _make_frontier_receipt,
-)
-from ..spurious_dragon import SpuriousDragonVM
-from ..spurious_dragon.computation import SpuriousDragonComputation
+from evm.vm.forks.spurious_dragon import SpuriousDragonVM
 
 from .constants import EIP649_BLOCK_REWARD
 from .headers import (
     create_byzantium_header_from_parent,
     configure_byzantium_header,
 )
-from .opcodes import BYZANTIUM_OPCODES
 from .blocks import ByzantiumBlock
-
-
-BYZANTIUM_PRECOMPILES = merge(
-    FRONTIER_PRECOMPILES,
-    {
-        force_bytes_to_address(b'\x05'): precompiles.modexp,
-        force_bytes_to_address(b'\x06'): precompiles.ecadd,
-        force_bytes_to_address(b'\x07'): precompiles.ecmul,
-        force_bytes_to_address(b'\x08'): precompiles.ecpairing,
-    },
-)
+from .computation import ByzantiumComputation
+from .vm_state import ByzantiumVMState
 
 
 def _byzantium_get_block_reward(block_number):
@@ -51,30 +24,15 @@ def _byzantium_get_uncle_reward(block_number, uncle):
     return (8 - block_number_delta) * EIP649_BLOCK_REWARD // 8
 
 
-def make_byzantium_receipt(vm, transaction, computation):
-    old_receipt = _make_frontier_receipt(vm, transaction, computation)
-    receipt = Receipt(
-        state_root=b'' if computation.is_error else b'\x01',
-        gas_used=old_receipt.gas_used,
-        logs=old_receipt.logs,
-    )
-    return receipt
-
-
 ByzantiumVM = SpuriousDragonVM.configure(
     name='ByzantiumVM',
-    # precompiles
-    _precompiles=BYZANTIUM_PRECOMPILES,
-    # opcodes
-    opcodes=BYZANTIUM_OPCODES,
-    # State
-    _computation_class=SpuriousDragonComputation,
-    # RLP
+    # classes
     _block_class=ByzantiumBlock,
+    _computation_class=ByzantiumComputation,
+    _state_class=ByzantiumVMState,
     # Methods
     create_header_from_parent=staticmethod(create_byzantium_header_from_parent),
     configure_header=configure_byzantium_header,
     get_block_reward=staticmethod(_byzantium_get_block_reward),
     get_uncle_reward=staticmethod(_byzantium_get_uncle_reward),
-    make_receipt=make_byzantium_receipt,
 )
