+++ evm/vm/forks/homestead/__init__.py
@@ -1,10 +1,8 @@
 from evm.chains.mainnet.constants import (
     DAO_FORK_BLOCK_NUMBER
 )
+from evm.vm.forks.frontier import FrontierVM
 
-from ..frontier import FrontierVM
-
-from .opcodes import HOMESTEAD_OPCODES
 from .blocks import HomesteadBlock
 from .computation import HomesteadComputation
 from .validation import validate_homestead_transaction
@@ -12,6 +10,7 @@ from .headers import (
     create_homestead_header_from_parent,
     configure_homestead_header,
 )
+from .vm_state import HomesteadVMState
 
 
 class MetaHomesteadVM(FrontierVM):
@@ -21,9 +20,10 @@ class MetaHomesteadVM(FrontierVM):
 
 HomesteadVM = MetaHomesteadVM.configure(
     name='HomesteadVM',
-    opcodes=HOMESTEAD_OPCODES,
+    # classes
     _block_class=HomesteadBlock,
     _computation_class=HomesteadComputation,
+    _state_class=HomesteadVMState,
     # method overrides
     validate_transaction=validate_homestead_transaction,
     create_header_from_parent=staticmethod(create_homestead_header_from_parent),
