+++ evm/vm/forks/homestead/computation.py
@@ -8,12 +8,22 @@ from evm.utils.hexadecimal import (
 from evm.utils.keccak import (
     keccak,
 )
-from ..frontier.computation import (
+from evm.vm.forks.frontier.computation import (
     FrontierComputation,
 )
 
+from .opcodes import HOMESTEAD_OPCODES
+
 
 class HomesteadComputation(FrontierComputation):
+    def __init__(self, vm_state, message):
+        super(HomesteadComputation, self).__init__(
+            vm_state,
+            message,
+        )
+        # Overwrite
+        self.opcodes = HOMESTEAD_OPCODES
+
     def apply_create_message(self):
         snapshot = self.vm_state.snapshot()
