+++ evm/vm/forks/homestead/vm_state.py
@@ -1,0 +1,5 @@
+from evm.vm.forks.frontier.vm_state import FrontierVMState
+
+
+class HomesteadVMState(FrontierVMState):
+    pass
