+++ evm/vm/forks/spurious_dragon/computation.py
@@ -8,13 +8,23 @@ from evm.utils.hexadecimal import (
 from evm.utils.keccak import (
     keccak,
 )
-from .constants import EIP170_CODE_SIZE_LIMIT
-from ..homestead.computation import (
+from evm.vm.forks.homestead.computation import (
     HomesteadComputation,
 )
 
+from .constants import EIP170_CODE_SIZE_LIMIT
+from .opcodes import SPURIOUS_DRAGON_OPCODES
+
 
 class SpuriousDragonComputation(HomesteadComputation):
+    def __init__(self, vm_state, message):
+        super(SpuriousDragonComputation, self).__init__(
+            vm_state,
+            message,
+        )
+        # Overwrite
+        self.opcodes = SPURIOUS_DRAGON_OPCODES
+
     def apply_create_message(self):
         snapshot = self.vm_state.snapshot()
