+++ evm/vm/forks/tangerine_whistle/vm_state.py
@@ -1,0 +1,5 @@
+from evm.vm.forks.homestead.vm_state import HomesteadVMState
+
+
+class TangerineWhistleVMState(HomesteadVMState):
+    pass
